Address review feedback: deduplicate conftest, add SPDX headers and comments
- Move get_slangpy_include_paths and get_artifact_path into test_utils.py
  to eliminate duplication and fragile importlib hack
- Simplify conftest.py to pytest markers only
- Add SPDX license headers to stress test Python and Slang files
- Add comment explaining CUDA-only CoopMat device selection
- Add comment explaining per-entry-point typealias duplication in CoopMat shader
