+++ tests/integration/slangpy/neural-slang-stress/test_mlp_only_baseline.py
@@ -1,3 +1,5 @@
+# SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+
 from __future__ import annotations
 
 import numpy as np
